requirements.txt
langchain
langchain-community
faiss-cpu
langchain-openai
openai
unstructured[docx]
sentence-transformers
python-dotenv
tiktoken 
Flask
waitress
redis==4.6.0
schedule